fix: change get product list by user
- 기존 secret은 의미를 알 수 없어 info로 라우터를 옮김
- 해당 라우터는 개별 사용자가 데이터를 조회하기 위함임
