+++ apis/product/app.js
@@ -5,7 +5,6 @@ import cors from 'cors';
 import indexRouter from './routes/index';
 import productRouter from './routes/product';
 import infoRouter from './routes/info';
-import secretRouter from './routes/secret';
 import etagGenerator from './routes/middleware/etag-generator';
 import { dbConnect } from './config';
 
@@ -24,7 +23,6 @@ app.get('*', etagGenerator);
 app.use('/', indexRouter);
 app.use('/info', infoRouter);
 app.use('/products', productRouter);
-app.use('/secrets', secretRouter);
 
 // eslint-disable-next-line no-unused-vars
 app.use((err, req, res, next) => {
